requirements.txt
torch
torchvision
pillow
pandas
opencv-python
torchsummary
matplotlib